moved list_installed into its own file
