+++ openmdao/utils/tests/test_cmdline.py
@@ -21,16 +21,17 @@ def _test_func_name(func, num, param):
 
 cmd_tests = [
     'openmdao call_tree openmdao.components.exec_comp.ExecComp.setup',
-    'openmdao summary {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao check {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao cite {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao iprof --no_browser {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao iprof_totals {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
+    'openmdao list_installed components commands nl_solvers lin_solvers drivers',
     'openmdao n2 --no_browser {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao n2 --no_browser {} -- -f bar'.format(os.path.join(scriptdir, 'circle_coloring_needs_args.py')),
     'openmdao partial_coloring {}'.format(os.path.join(scriptdir, 'circle_coloring_dynpartials.py')),
     'openmdao scaffold -b ExplicitComponent -c Foo',
     'openmdao scaffold -b ImplicitComponent -c Foo',
+    'openmdao summary {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao total_coloring {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao trace {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
     'openmdao tree -c {}'.format(os.path.join(scriptdir, 'circle_opt.py')),
@@ -66,7 +67,7 @@ class CmdlineTestCase(unittest.TestCase):
 
     @parameterized.expand(mem_cmd_tests, name_func=_test_func_name)
     @unittest.skipIf(psutil is None, 'psutil must be installed to run mem commands')
-    def test_cmd(self, cmd):
+    def test_cmd2(self, cmd):
         # this only tests that a given command line tool returns a 0 return code. It doesn't
         # check the expected output at all.  The underlying functions that implement the
         # commands should be tested seperately.
